Adding a second 2k11 prototype build to the legacy content archive.
This build was done in the same year as the previous one but is much different in almost every way.  More robots, a few more fields, an actual campaign with a post-game, and even a secret character.  This one, while very very slow, is definitely worth checking out at one point or another.